+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+# sankeymien v0.0.2 (2026-06-17)
+
+## Add
+
+- `abundance_threshold` to filter organism abundance.
+- `relative` to use relative abundance instead of absolute.
+- tests for `handle_input` and `generate_node_edges`.
+
+## Modify
+
+- Refactor how node numbers are found to avoid several issues.
+
 # sankeymien v0.0.1 (2026-06-17)
 
 First draft of package.